apps/brokerservice/requirements.txt
grpcio==1.19.0
kerberos==1.3.0
Authlib==0.10
redis==2.10.6
structlog==19.1.0
python-redis-lock==3.2.0
google-auth-oauthlib==0.2.0
google-cloud-datastore==1.7.3
google-cloud-kms==0.2.1